fixed header and added hamburger. content switches properly to different pages. content still needs minor adjustment to make readable
