+++ package.json
@@ -14,6 +14,7 @@
     "gh-pages": "^6.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.4.0",
     "react-router": "^6.8.x",
     "react-router-dom": "^6.8.x",
     "react-scripts": "^5.0.1",
